Updated makefile and serial TPFA solver.
Fixed makefile to add Trilinos libraries links. Started to implement a
completely serial version of the TPFA solver for future use in the
wraper for Python. This version is broke, meaning I'm still working on
changes to improve the assembly time. Apparently the assembly increases greatly in a full serial version.

Some possible fixes and notes to future work:
 - Remove all data access to the mesh from the main loop.
 - Check how multiple values properties are retrieved by MOAB (in a
   sequence or in an array of arrays).