Global ID issue and matrix assembly
This commit contains a version of the TPFA solver with issue concerning
the volume's ids. A mesh created by MOAB does not supply a default value
for global ID. Therefore, it must be created and assigned to all
volumes.

The main problem is that this new ID must be somehow equivalent to the
ones used by Epetra to fill the matrix so that the correct properties be
used to calculate the coeficient.